multiy point light sources in shading process

+++ src/screen.h
@@ -10,9 +10,9 @@
 
 
 struct Color {
-	uint8_t r;
-	uint8_t b;
-	uint8_t g;
+	double r;
+	double b;
+	double g;
 };
 
 class Screen {
